Move UInt16Tests
