+++ Tests/Batch1/SimpleTypes/UInt16Tests.cs
@@ -1,5 +1,6 @@
 ﻿using Bridge.Test.NUnit;
 using System;
+using System.Globalization;
 
 namespace Bridge.ClientTest.SimpleTypes
 {
@@ -8,19 +9,28 @@ namespace Bridge.ClientTest.SimpleTypes
     public class UInt16Tests
     {
         [Test]
-        public void TypePropertiesAreCorrect()
+        public void TypePropertiesAreCorrect_SPI_1717()
         {
             Assert.True((object)(ushort)0 is ushort);
             Assert.False((object)0.5 is ushort);
             Assert.False((object)-1 is ushort);
             Assert.False((object)65536 is ushort);
             Assert.AreEqual("System.UInt16", typeof(ushort).FullName);
-
+            Assert.False(typeof(ushort).IsClass);
+            Assert.True(typeof(IComparable<ushort>).IsAssignableFrom(typeof(ushort)));
+            Assert.True(typeof(IEquatable<ushort>).IsAssignableFrom(typeof(ushort)));
+            Assert.True(typeof(IFormattable).IsAssignableFrom(typeof(ushort)));
             object s = (ushort)0;
             Assert.True(s is ushort);
             Assert.True(s is IComparable<ushort>);
             Assert.True(s is IEquatable<ushort>);
             Assert.True(s is IFormattable);
+
+            var interfaces = typeof(ushort).GetInterfaces();
+            Assert.AreEqual(4, interfaces.Length);
+            Assert.True(interfaces.Contains(typeof(IComparable<ushort>)));
+            Assert.True(interfaces.Contains(typeof(IEquatable<ushort>)));
+            Assert.True(interfaces.Contains(typeof(IFormattable)));
         }
 
         [Test]
@@ -47,17 +57,29 @@ namespace Bridge.ClientTest.SimpleTypes
 
             checked
             {
-                Assert.Throws(() => { var b = (ushort)i1; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (ushort)i1;
+                }, "-1 checked");
                 Assert.AreStrictEqual(0, (ushort)i2, "0 checked");
                 Assert.AreStrictEqual(234, (ushort)i3, "234 checked");
                 Assert.AreStrictEqual(65535, (ushort)i4, "65535 checked");
-                Assert.Throws(() => { var b = (ushort)i5; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (ushort)i5;
+                }, "65536 checked");
 
-                Assert.Throws(() => { var b = (ushort?)ni1; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (ushort?)ni1;
+                }, "nullable -1 checked");
                 Assert.AreStrictEqual(0, (ushort?)ni2, "nullable 0 checked");
                 Assert.AreStrictEqual(234, (ushort?)ni3, "nullable 234 checked");
                 Assert.AreStrictEqual(65535, (ushort?)ni4, "nullable 65535 checked");
-                Assert.Throws(() => { var b = (ushort?)ni5; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (ushort?)ni5;
+                }, "nullable 65536 checked");
                 Assert.AreStrictEqual(null, (ushort?)ni6, "null checked");
             }
         }
@@ -99,13 +121,32 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
-        public void IFormattableToStringWorks()
+        public void ToStringWithFormatWorks()
         {
             Assert.AreEqual("123", ((ushort)0x123).ToString("x"));
         }
 
         [Test]
-        public void TryParseWorks()
+        public void ToStringWithFormatAndProviderWorks()
+        {
+            Assert.AreEqual("123", ((ushort)0x123).ToString("x", CultureInfo.InvariantCulture));
+        }
+
+        [Test]
+        public void IFormattableToStringWorks()
+        {
+            Assert.AreEqual("123", ((IFormattable)((ushort)0x123)).ToString("x", CultureInfo.InvariantCulture));
+        }
+
+        // Not C# API
+        //[Test]
+        //public void LocaleFormatWorks()
+        //{
+        //    Assert.AreEqual(((ushort)0x123).LocaleFormat("x"), "123");
+        //}
+
+        [Test]
+        public void TryParseWorks_SPI_1592()
         {
             ushort numberResult;
             bool result = ushort.TryParse("23445", out numberResult);
@@ -130,7 +171,9 @@ namespace Bridge.ClientTest.SimpleTypes
 
             result = ushort.TryParse("-1", out numberResult);
             Assert.False(result);
+            // #1592
             Assert.AreEqual(-1, numberResult);
+            //Assert.AreEqual(0, numberResult);
 
             result = ushort.TryParse("2.5", out numberResult);
             Assert.False(result);
@@ -141,12 +184,12 @@ namespace Bridge.ClientTest.SimpleTypes
         public void ParseWorks()
         {
             Assert.AreEqual(23445, ushort.Parse("23445"));
-            Assert.Throws(() => ushort.Parse(""));
-            Assert.Throws(() => ushort.Parse(null));
-            Assert.Throws(() => ushort.Parse("notanumber"));
-            Assert.Throws(() => ushort.Parse("65536"));
-            Assert.Throws(() => ushort.Parse("-1"));
-            Assert.Throws(() => ushort.Parse("2.5"));
+            Assert.Throws<FormatException>(() => ushort.Parse(""));
+            Assert.Throws<ArgumentNullException>(() => ushort.Parse(null));
+            Assert.Throws<FormatException>(() => ushort.Parse("notanumber"));
+            Assert.Throws<OverflowException>(() => ushort.Parse("65536"));
+            Assert.Throws<OverflowException>(() => ushort.Parse("-1"));
+            Assert.Throws<FormatException>(() => ushort.Parse("2.5"));
         }
 
         [Test]
